+++ src/components/PartsOfPlanTable/PartsOfPlanTable.tsx
@@ -9,6 +9,8 @@ import './PartsOfPlanTable.css';
 import {GoBack} from "../GoBack/GoBack";
 import {MoonLoader} from "react-spinners";
 import {ReusableModal} from "../ReusableModal/ReusableModal";
+import {isDemoEnabled} from "../hooks/env";
+import {DemoSign} from "../DemoSign/DemoSign";
 
 export const PartsOfPlanTable = () => {
 
@@ -18,11 +20,14 @@ export const PartsOfPlanTable = () => {
     const [partToDeleteId, setPartToDeleteId] = useState(null);
     const [informationModalIsOpen, setInformationModalIsOpen] = useState<boolean>(false);
     const [trainingPlanName, setTrainingPlanName] = useState('');
+    const [demoModalIsOpen, setDemoModalIsOpen] = useState<boolean>(false);
     const [isLoading, setIsLoading] = useState(true);
 
     const params = useParams();
 
     const text = 'Czy na pewno chcesz usunąć tę część planu? Spowoduje to także usunięcie wszystkich ćwiczeń przypisanych do tej części planu';
+
+    const demoText = 'To jest wersja demo aplikacji "Gym Training Planner". Nie można w niej dodawać, edytować ani usuwać wybranych elementów.'
 
     const textInformation = 'Należy podać nazwę części planu!'
 
@@ -72,34 +77,49 @@ export const PartsOfPlanTable = () => {
         setInformationModalIsOpen(false);
     };
 
-    const addPartOfPlan = async (values: PartOfPlanEntity) => {
-        fetch(`${apiUrl}/api/add-plan/list?slug=${params.slug}`, {
-            method: 'GET',
-        })
-            .then(r => r.json())
-            .then(async (plan) => {
-                if (!plan || plan.length === 0) {
-                    console.log('Brak planu.')
-                } else {
-
-                    const res = await fetch(`${apiUrl}/api/add-part/plans`, {
-                        method: 'POST',
-                        headers: {
-                            'Content-Type': 'application/json',
-                        },
-                        body: JSON.stringify({...values, planId: plan[0].id}),
-                    })
-
-                    const data = await res.json();
-
-                    setPartsOfPlanList(list => [...list, data]);
-
-                }
-            })
+    const closeDemoModal = () => {
+        setDemoModalIsOpen(false);
     };
 
-    const editPartOfPlan = async (values: PartOfPlanEntity) => {
+    const addPartOfPlan = async (values: PartOfPlanEntity) => {
+        if (isDemoEnabled()) {
+            setDemoModalIsOpen(true);
+        } else if (values.name) {
+            fetch(`${apiUrl}/api/add-plan/list?slug=${params.slug}`, {
+                method: 'GET',
+            })
+                .then(r => r.json())
+                .then(async (plan) => {
+                    if (!plan || plan.length === 0) {
+                        console.log('Brak planu.')
+                    } else {
+
+                        const res = await fetch(`${apiUrl}/api/add-part/plans`, {
+                            method: 'POST',
+                            headers: {
+                                'Content-Type': 'application/json',
+                            },
+                            body: JSON.stringify({...values, planId: plan[0].id}),
+                        })
+
+                        const data = await res.json();
+
+                        setPartsOfPlanList(list => [...list, data]);
+
+                    }
+                })
+        } else {
+            setInformationModalIsOpen(true);
+        }
+    };
+
+    const editPartOfPlan = async (values: PartOfPlanEntity, reset: () => void) => {
         setIsEdited(false);
+
+        if (isDemoEnabled()) {
+            setDemoModalIsOpen(true);
+            reset();
+        } else if (values.name) {
 
         const res = await fetch(`${apiUrl}/api/add-part/plans/${values.id}`, {
             method: 'PUT',
@@ -117,6 +137,10 @@ export const PartsOfPlanTable = () => {
 
         return await res.json();
 
+        } else {
+            setInformationModalIsOpen(true);
+            reset();
+        }
     };
 
     const handleUpdatePartOfPlan = (updatedPart: PartOfPlanEntity) => {
@@ -128,25 +152,34 @@ export const PartsOfPlanTable = () => {
     };
 
     const handleDeletePart = async (partId: any) => {
-        setConfirmDeletePart(true);
-        setPartToDeleteId(partId);
+        if (isDemoEnabled()) {
+            setDemoModalIsOpen(true);
+            setPartToDeleteId(partId);
+        } else {
+            setConfirmDeletePart(true);
+            setPartToDeleteId(partId);
+        }
     };
 
     const handleConfirmDelete = async () => {
-        const res = await fetch(
-            `${apiUrl}/api/add-part/plans/${partToDeleteId}`,
-            {method: "DELETE"}
-        )
-        if ([400, 500].includes(res.status)) {
-            const error = await res.json();
-            alert(`Wystąpił błąd: ${error.message}`);
-            return;
-        }
-        setPartsOfPlanList((partsOfPlanList) =>
-            partsOfPlanList.filter((part) => part.id !== partToDeleteId)
-        );
-        setConfirmDeletePart(false);
-
+        if (isDemoEnabled()) {
+            closeDemoModal();
+        } else {
+            const res = await fetch(
+                `${apiUrl}/api/add-part/plans/${partToDeleteId}`,
+                {method: "DELETE"}
+            )
+            if ([400, 500].includes(res.status)) {
+                const error = await res.json();
+                alert(`Wystąpił błąd: ${error.message}`);
+                return;
+            }
+            setPartsOfPlanList((partsOfPlanList) =>
+                partsOfPlanList.filter((part) => part.id !== partToDeleteId)
+            );
+            setConfirmDeletePart(false);
+            setPartToDeleteId(null);
+        };
     };
 
     const handleCancelDelete = () => {
@@ -158,7 +191,7 @@ export const PartsOfPlanTable = () => {
         return (
             <div className="spinner_container">
                 <div className="div_loading">Ładowanie częsci planu..</div>
-                <MoonLoader speedMultiplier={0.5} color="#9fc3f870" />
+                <MoonLoader speedMultiplier={0.5} color="#9fc3f870"/>
             </div>
         );
     }
@@ -169,6 +202,7 @@ export const PartsOfPlanTable = () => {
                 <h1 className="main-h1"><TbHeartbeat/> Gym Training Planner</h1>
             </IconContext.Provider>
             <div className="main-plan">
+                <DemoSign/>
                 <table className="main-table">
 
                     <thead>
@@ -196,18 +230,14 @@ export const PartsOfPlanTable = () => {
                                 name: '',
                             }}
                             onSubmit={async (values, reset) => {
-                                if (values.name) {
-                                    await addPartOfPlan(values);
-                                    reset();
-                                } else {
-                                    setInformationModalIsOpen(true);
-                                }
+                                await addPartOfPlan(values);
+                                reset();
                             }}
                             actionType={Status.Add}
                         />
                         <td className="td-progression-rules">
                             <IconContext.Provider value={{className: 'react-icons-progression'}}>
-                                <Link to={`/rules/${params.slug}`}><TbStairsUp /></Link>
+                                <Link to={`/rules/${params.slug}`}><TbStairsUp/></Link>
                             </IconContext.Provider>
                         </td>
                     </tr>
@@ -221,14 +251,9 @@ export const PartsOfPlanTable = () => {
                             </td>
                             <PartsOfPlanForm
                                 initialValues={part}
-                                onSubmit={async (values) => {
-                                    if (values.name) {
-                                        await editPartOfPlan(values);
-                                        await handleUpdatePartOfPlan(values);
-                                    } else {
-                                        setInformationModalIsOpen(true);
-                                        values.name = part.name;
-                                    }
+                                onSubmit={async (values, reset) => {
+                                    await editPartOfPlan(values, reset);
+                                    await handleUpdatePartOfPlan(values);
                                 }}
                                 actionType={Status.Save}
                                 isEdited={isEdited}
@@ -263,6 +288,16 @@ export const PartsOfPlanTable = () => {
                 icon={TbAlertTriangle}
                 confirmText="Rozumiem"
             />
+            {demoModalIsOpen && (
+                <ReusableModal
+                    isOpen={demoModalIsOpen}
+                    onRequestClose={closeDemoModal}
+                    onConfirm={closeDemoModal}
+                    text={demoText}
+                    icon={TbAlertTriangle}
+                    confirmText="OK"
+                />
+            )}
         </div>
     )
 }
